Prepare the next release.

+++ tests/test_python.py
@@ -2,8 +2,7 @@ import pytest
 from hypothesis import given
 
 from tests.strategies import byte_strategy, bytes_strategy
-from xor_cipher.core import cyclic_xor, cyclic_xor_in_place, xor, xor_in_place
-from xor_cipher.typing import byte
+from xor_cipher.python import cyclic_xor, cyclic_xor_in_place, xor, xor_in_place
 
 EXPECTED = b"Hello, world!"
 
@@ -46,7 +45,7 @@ def given_cyclic_key() -> bytes:
 
 
 @given(bytes_strategy, byte_strategy)
-def test_xor(data: bytes, key: byte) -> None:
+def test_xor(data: bytes, key: int) -> None:
     assert xor(xor(data, key), key) == data
 
 
@@ -61,7 +60,7 @@ def test_cyclic_xor(data: bytes, key: bytes) -> None:
     assert cyclic_xor(cyclic_xor(data, key), key) == data
 
 
-def test_xor_in_place_expected(given_data: bytes, given_key: byte, expected_data: bytes) -> None:
+def test_xor_in_place_expected(given_data: bytes, given_key: int, expected_data: bytes) -> None:
     array = bytearray(given_data)
 
     xor_in_place(array, given_key)
@@ -70,7 +69,7 @@ def test_xor_in_place_expected(given_data: bytes, given_key: byte, expected_data
 
 
 @given(bytes_strategy, byte_strategy)
-def test_xor_in_place(data: bytes, key: byte) -> None:
+def test_xor_in_place(data: bytes, key: int) -> None:
     array = bytearray(data)
 
     xor_in_place(array, key)
